Move custom views to Views namespace
