+++ visitz/Visitz/Views/CaseloadItemView.xaml
@@ -1,10 +1,10 @@
 <?xml version="1.0" encoding="utf-8" ?>
 <ContentView xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-             xmlns:controls="clr-namespace:Visitz.Controls"
+             xmlns:views="clr-namespace:Visitz.Views"
              xmlns:models="clr-namespace:Visitz.Models"
              xmlns:viewmodels="clr-namespace:Visitz.ViewModels"
-             x:Class="Visitz.Controls.CaseloadItemView"
+             x:Class="Visitz.Views.CaseloadItemView"
              x:DataType="models:CaseloadItem">
     
     <VerticalStackLayout Padding="20,0,20,0" Margin="0,10,0,0">
