+++ visitz/Visitz/Views/CaseloadItemView.xaml.cs
@@ -1,4 +1,4 @@
-namespace Visitz.Controls;
+namespace Visitz.Views;
 
 public partial class CaseloadItemView : ContentView
 {
