+++ visitz/Visitz/Views/ContactItemView.xaml
@@ -1,9 +1,9 @@
 <?xml version="1.0" encoding="utf-8" ?>
 <ContentView xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-             xmlns:controls="clr-namespace:Visitz.Controls"
+             xmlns:views="clr-namespace:Visitz.Views"
              xmlns:models="clr-namespace:Visitz.Models"
-             x:Class="Visitz.Controls.ContactItemView"
+             x:Class="Visitz.Views.ContactItemView"
              x:DataType="models:FamilyMember">
 
     <VerticalStackLayout>
