+++ visitz/Visitz/Views/FormItemView.xaml
@@ -1,9 +1,9 @@
 <?xml version="1.0" encoding="utf-8" ?>
 <ContentView xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-             xmlns:controls="clr-namespace:Visitz.Controls"
-             x:Class="Visitz.Controls.FormItemView"
-             x:DataType="controls:FormItemView"
+             xmlns:views="clr-namespace:Visitz.Views"
+             x:Class="Visitz.Views.FormItemView"
+             x:DataType="views:FormItemView"
              x:Name="this">
 
     <VerticalStackLayout BindingContext="{x:Reference this}">
