+++ visitz/Visitz/Visitz.csproj
@@ -154,7 +154,7 @@
 	  <Compile Update="Controls\FormItemViewView.xaml.cs">
 	    <DependentUpon>FormItemViewView.xaml</DependentUpon>
 	  </Compile>
-	  <Compile Update="Controls\CaseloadItemView.xaml.cs">
+	  <Compile Update="Views\CaseloadItemView.xaml.cs">
 	    <DependentUpon>CaseloadItemView.xaml</DependentUpon>
 	  </Compile>
 	  <Compile Update="Pages\CaseloadItemDetailsPage.xaml.cs">
@@ -188,13 +188,13 @@
 	  </Compile>
 	</ItemGroup>
 	<ItemGroup>
-	  <MauiXaml Update="Controls\ContactItemView.xaml">
+	  <MauiXaml Update="Views\ContactItemView.xaml">
 	    <Generator>MSBuild:Compile</Generator>
 	  </MauiXaml>
 	  <MauiXaml Update="Controls\FormItemViewView.xaml">
 	    <Generator>MSBuild:Compile</Generator>
 	  </MauiXaml>
-	  <MauiXaml Update="Controls\CaseloadItemView.xaml">
+	  <MauiXaml Update="Views\CaseloadItemView.xaml">
 	    <Generator>MSBuild:Compile</Generator>
 	  </MauiXaml>
 	  <MauiXaml Update="Pages\DebugOptionsPage.xaml">
